Formatted a home page
Added a lot of extra material to the .css file in order to format many more things, both for topnavigation-bar, sidenavigation, image-formatting, and more. Experimenting with different things, so it's not all used equally much, but the home page now has two different looks that we can build on further. Also, the .css has two different styles for h1 headers. One of them is just commented out, but they still both can be used to format the header in different ways